+++ src/Castle.Core/Core/Internal/ILockHolder.cs
@@ -15,7 +15,10 @@
 namespace Castle.Core.Internal
 {
 	using System;
+	using System.ComponentModel;
 
+	[Obsolete("Consider using `System.Threading.ReaderWriterLockSlim` instead of `Lock` and related types.")] // TODO: Remove this type.
+	[EditorBrowsable(EditorBrowsableState.Never)]
 	public interface ILockHolder:IDisposable
 	{
 		bool LockAcquired { get; }