+++ src/Castle.Core/DynamicProxy/Serialization/CacheMappingsAttribute.cs
@@ -18,6 +18,7 @@ namespace Castle.DynamicProxy.Serialization
 {
 	using System;
 	using System.Collections.Generic;
+	using System.ComponentModel;
 	using System.IO;
 	using System.Reflection;
 	using System.Reflection.Emit;
@@ -47,6 +48,8 @@ namespace Castle.DynamicProxy.Serialization
 			get { return serializedCacheMappings; }
 		}
 
+		[Obsolete("Exposes a component that is intended for internal use only.")] // TODO: Redeclare this method as `internal`.
+		[EditorBrowsable(EditorBrowsableState.Never)]
 		public Dictionary<CacheKey, string> GetDeserializedMappings()
 		{
 			using (var stream = new MemoryStream(SerializedCacheMappings))
@@ -56,6 +59,8 @@ namespace Castle.DynamicProxy.Serialization
 			}
 		}
 
+		[Obsolete("Exposes a component that is intended for internal use only.")] // TODO: Redeclare this method as `internal`.
+		[EditorBrowsable(EditorBrowsableState.Never)]
 		public static void ApplyTo(AssemblyBuilder assemblyBuilder, Dictionary<CacheKey, string> mappings)
 		{
 			using (var stream = new MemoryStream())